Password label contraint now derives value from containing view
This change was made based on some feedback about scalability regarding the bottom constraint value of the pw labels when they are being set in the VC. Previously, they were hard coded in the animation's completetion block AND the storyboard. We now derive the value from the container view's outlet.