(Release) Cut release for v1.2 (#10)
* Password label contraint now derives value from containing view

This change was made based on some feedback about scalability regarding the bottom constraint value of the pw labels when they are being set in the VC. Previously, they were hard coded in the animation's completetion block AND the storyboard. We now derive the value from the container view's outlet.

* Feature/add more share options (#8)

### Summary
Addition of a share action sheet. The user now has the ability to share a generated password via multiple share actions by long pressing the password label.

### Test Plan
- Generate password
- Long press the password label
- Select a share action

* Refactored constants

* Renaming

* Unit tests for password characters

* More test updates

* Cleanup

* TODO Comment added and cleanup

* Removed whitespace

* (Task) | Added New Share Functionality (#9)

### Summary
This adds new share functionality which makes use of UIActionSheet. Other minor cleanup and organization included.

### Test Steps
- Generate new password
- Long-press on the newly generated password

You should be presented with options for sharing the password.

* Implement Travis CI

This adds support for Travis CI

* Travis config update

Typo in the yml file

* Revert "Travis config update"

This reverts commit 13dae29ffb07811f4c27d28be17e70e5d16c7364.

* I need more coffee

* Add Executable for Scheme

* CocoaPods

* Update .gitignore for cocoapods

* Remove CocoaPod

* View model cleanup and added unit tests

* Update travis config

The workspace was deleted when CocoaPods was deleted. This was causing build failures in Travis.